testing cross plaform tpm open func
